56: Cart Page Ui Changed To New One Accordign to the PDF✅ 57: Calculate Price For Cart Items Helper Updated and added items_total whihc just gives Only Total items Cost,✅
58: Checkout Page Redesgined ANd Ui Completed According to The PDF ✅
59: Lucid React Icon Library Added ✅
60: New 3 Columns Added In Orders Table -> (
    "points_used" => For Tracking How much Amount it  Paid For This Order Using Points
    "full_amount" => Tracking THe Full Order Amount
    "secondary_payment_method" To Track The 2nd Payment Method Used For the order when Uisng Points payment
    )✅

 61: Accross the Site Amount COlumn For order Changed To Full_amount where i was using amount column i replaced it with full_amount of order column
     And Now Amount Column For ORders Using as Remeaning Amount of Order Accross The Site✅
 62: Checkout Full Flow Rewritten According to new Payment Option Logics And Points + Payment Method Logics as Discussed with Joseph ✅
 63: (Cart Items Logic Updates Before we were showing cart items on checkout and on cart But In Checkout User Only can view the Items
     But Now THe new design showcased user on checkout page can also edit delete The Items So Cart Items Shown There in checkout page as well
     But there were was one Critical Issue is that when updating cart item from checkout page if user comes from  buy_now button by clicking directly
     from product page to checkout, so there wasnt any cart items in the cart becasue when user directly tries to purchase any product By clicking
     on Buynow button it simply creates a Session and Stores as That cart items so our update logic for the cart is asks for actual cart item but
     there wasnt any so The Logic To Handle The Cart Item + Session Cart Items If BuyNow Implemented ensuring the Proper Calculation and Quantity Updates
     And Same Applies For The Addon Items) ✅

64: Critical Issue Fixed: From Order view page before anyone can view anyones Order it was way critical Now i have identified that issue so i have Restricted
    Order View Only That order user can view which Belogns To Him Personally ✅

65: Social Media Helper Added InCheckout Page BEfore we were showign Buttons Diretcly below the Order Summary card
    Which was breaking the UX so now it SHows as helper Iocn right corner From Bottom of the Site And when clickign on it
    The Modal popups On the Right side and Shows the Social Media Items So when clicign on it it will diretcly take user to the Sites social media Platform (If Configured Proper URL From Backend Meta Configuration) ✅

66: In Now Payment invoice Status Check And Auto marking Order Failed if not paid Command Updated And Added Reward points Rollbacking If Used For the
    Specific Order ✅

 67: All Notifications Related To Order Updated And Added New Column From order table full_amount replacing amoutn where i was showing ful_amount of the order in the Notification and added 2 Colums Remeaning Amount And Full amount to differenciate Properly On appropriate Places in notifications ✅

 68: Spatiotemporal and Global Filter Modal backdrop Blur Removed To Make  consistency Accross the site ✅
 69:  Order Table Columns Updated in backend and added 3 more columns in it to get Full admin View (Points used, full amount, amount ),
 70:  ORder View In Backend Order Summary Updated and added 3 new rows one shows full amount and one shows Remeanign amount and one shows Points payment ✅,

 71: Order Refund Backend Updated now it proerly stores The Refund Full amount before it was using the amount column from order which now is using as Remeaning amount ✅

+++ app/Repositories/Cart/Repository/CartRepository.php
@@ -305,37 +305,56 @@ class CartRepository implements ICartRepository
 
             $item_type = $request->input('type');
             $item_id = $request->integer('item_id');
+            $page = $request->input('page');
 
             if ($item_type === 'smartphone') {
-                $item = $this->cart->where('customer_id', $customer->id)->where('id', $item_id)->with(['smartphone', 'smartphoneAddonItems'])->first();
-                if (empty($item)) {
-                    throw new Exception($this->trans->get('Something Went Wrong While Removing Product From Cart'));
-                }
 
-                DB::transaction(function () use ($item) {
-                    if ($item->smartphone) {
-                        $item->smartphoneAddonItems()
-                            ->delete();
-                    }
-                    $deleted = $item->delete();
-                    if (! $deleted) {
+                if ($page === 'cart') {
+                    $item = $this->cart->where('customer_id', $customer->id)->where('id', $item_id)->with(['smartphone', 'smartphoneAddonItems'])->first();
+                    if (empty($item)) {
                         throw new Exception($this->trans->get('Something Went Wrong While Removing Product From Cart'));
                     }
-                });
 
-                $all_cart_items = $this->cart->where('customer_id', $customer->id)->with(['smartphone', 'smartphone.selling_info', 'smartphone.selling_info.shipping_fee', 'smartphone.selling_info.import_tax'])->get();
+                    DB::transaction(function () use ($item) {
+                        if ($item->smartphone) {
+                            $item->smartphoneAddonItems()
+                                ->delete();
+                        }
+                        $deleted = $item->delete();
+                        if (! $deleted) {
+                            throw new Exception($this->trans->get('Something Went Wrong While Removing Product From Cart'));
+                        }
+                    });
 
-                $addon_cart_items = $this->smartphone_cart_addon
-                    ->where('customer_id', $customer->id)
-                    ->whereIn('smartphone_id', $all_cart_items->pluck('smartphone_id')->toArray())
-                    ->get();
+                    $all_cart_items = $this->cart->where('customer_id', $customer->id)->with(['smartphone', 'smartphone.selling_info', 'smartphone.selling_info.shipping_fee', 'smartphone.selling_info.import_tax'])->get();
 
-                $total_summary = $this->price_calculator->calculate($all_cart_items, $addon_cart_items);
+                    $addon_cart_items = $this->smartphone_cart_addon
+                        ->where('customer_id', $customer->id)
+                        ->whereIn('smartphone_id', $all_cart_items->pluck('smartphone_id')->toArray())
+                        ->get();
+
+                    $total_summary = $this->price_calculator->calculate($all_cart_items, $addon_cart_items);
+
+                    return [
+                        'status' => true,
+                        'message' => $this->trans->get('Removed Succesfully From Cart'),
+                        'total_summary' => $total_summary,
+                    ];
+                }
+
+                $sessionKey = 'single_product_checkout_'.$user->id;
+                $sessionData = session()->get($sessionKey);
+
+                if (empty($sessionData) || empty($sessionData['cart_items'])) {
+                    throw new Exception('Invalid checkout session');
+                }
+
+                session()->forget($sessionKey);
 
                 return [
                     'status' => true,
-                    'message' => $this->trans->get('Removed Succesfully From Cart'),
-                    'total_summary' => $total_summary,
+                    'message' => 'Item Removed Successfully',
+                    'total_summary' => [],
                 ];
 
             }
@@ -378,15 +397,61 @@ class CartRepository implements ICartRepository
                 throw new Exception($this->trans->get('Only Customers Can Remove items From Cart'));
             }
 
+            $page = $request->input('page');
             $item_type = $request->input('type');
             $item_id = $request->integer('item_id');
             $quantity = $request->integer('quantity');
 
-            $cart_item = $this->cart
-                ->where('customer_id', $customer->id)
-                ->where('id', $item_id)
-                ->with(['smartphone', 'smartphone.selling_info', 'smartphone.selling_info.shipping_fee', 'smartphone.selling_info.import_tax'])
-                ->get();
+            $cart_item = null;
+            if ($page === 'cart') {
+                $cart_item = $this->cart
+                    ->where('customer_id', $customer->id)
+                    ->where('id', $item_id)
+                    ->with(['smartphone', 'smartphone.selling_info', 'smartphone.selling_info.shipping_fee', 'smartphone.selling_info.import_tax'])
+                    ->get();
+            } else {
+
+                $sessionKey = 'single_product_checkout_'.$user->id;
+                $sessionData = session()->get($sessionKey);
+
+                if (empty($sessionData) || empty($sessionData['cart_items'])) {
+                    throw new Exception($this->trans->get('Invalid Buy Now Session'));
+                }
+
+                /** @var \Illuminate\Support\Collection $cartItems */
+                $cartItems = $sessionData['cart_items'];
+
+                /** @var \App\Models\CartItem $cartItem */
+                $cartItem = $cartItems->first();
+
+                if (! $cartItem) {
+                    throw new Exception($this->trans->get('Product not found in Buy Now session'));
+                }
+
+                $cartItem->quantity = $quantity;
+                $cartItem->total_price = bcmul(
+                    $cartItem->unit_price,
+                    $quantity,
+                    2
+                );
+
+                $total_summary = $this->price_calculator->calculate(
+                    $cartItems,
+                    $sessionData['addon_items'] ?? collect()
+                );
+
+                session()->put($sessionKey, [
+                    ...$sessionData,
+                    'cart_items' => $cartItems,
+                    'total_summary' => $total_summary,
+                ]);
+
+                return [
+                    'status' => true,
+                    'message' => $this->trans->get('Product Updated Succesfully'),
+                    'total_summary' => $total_summary,
+                ];
+            }
 
             if ($cart_item->isEmpty()) {
                 throw new Exception($this->trans->get('Wrong Product Selected Please Select Valid Product'));
@@ -451,8 +516,56 @@ class CartRepository implements ICartRepository
             $item_id = $request->integer('item_id');
             $quantity = $request->integer('quantity');
 
-            $cart_item = $this->smartphone_cart_addon->where('customer_id', $customer->id)->where('id', $item_id)->first();
+            $page = $request->input('page');
 
+            $cart_item = null;
+
+            if ($page === 'cart') {
+                $cart_item = $this->smartphone_cart_addon->where('customer_id', $customer->id)->where('id', $item_id)->first();
+            } else {
+
+                $sessionKey = 'single_product_checkout_'.$user->id;
+                $sessionData = $request->session()->get($sessionKey);
+
+                if (empty($sessionData) || empty($sessionData['addon_items'])) {
+                    throw new Exception('Invalid Buy Now Session');
+                }
+
+                /** @var \Illuminate\Support\Collection $addonItems */
+                $addonItems = $sessionData['addon_items'];
+
+                /** @var \App\Models\SmartphoneCartAddon $addonItem */
+                $addonItem = $addonItems->firstWhere('addon_id', $item_id);
+
+                // dd($addonItem);
+                if (! $addonItem) {
+                    throw new Exception('Addon not found in session');
+                }
+
+                $addonItem->quantity = $quantity;
+                $addonItem->total_price = bcmul(
+                    $addonItem->unit_price,
+                    $request->quantity,
+                    2
+                );
+
+                $total_summary = $this->price_calculator->calculate(
+                    $sessionData['cart_items'],
+                    $addonItems
+                );
+
+                session()->put($sessionKey, [
+                    ...$sessionData,
+                    'addon_items' => $addonItems,
+                    'total_summary' => $total_summary,
+                ]);
+
+                return [
+                    'status' => true,
+                    'message' => 'Addon Updated Successfully',
+                    'total_summary' => $total_summary,
+                ];
+            }
             if (empty($cart_item)) {
                 throw new Exception($this->trans->get('Wrong Product Selected Please Select Valid Product'));
             }
@@ -512,31 +625,79 @@ class CartRepository implements ICartRepository
             }
 
             $item_id = $request->integer('item_id');
+            $page = $request->input('page');
 
-            $item = $this->smartphone_cart_addon->where('customer_id', $customer->id)->where('id', $item_id)->first();
-            if (empty($item)) {
-                throw new Exception($this->trans->get('Something Went Wrong While Removing Addon From Cart'));
+            if ($page === 'cart') {
+                $item = $this->smartphone_cart_addon->where('customer_id', $customer->id)->where('id', $item_id)->first();
+                if (empty($item)) {
+                    throw new Exception($this->trans->get('Something Went Wrong While Removing Addon From Cart'));
+                }
+
+                $deleted = $item->delete();
+                if (! $deleted) {
+                    throw new Exception($this->trans->get('Something Went Wrong While Removing Addon From Cart'));
+                }
+
+                $all_cart_items = $this->cart->where('customer_id', $customer->id)->with(['smartphone', 'smartphone.selling_info', 'smartphone.selling_info.shipping_fee', 'smartphone.selling_info.import_tax'])->get();
+
+                $addon_cart_items = $this->smartphone_cart_addon
+                    ->where('customer_id', $customer->id)
+                    ->whereIn('smartphone_id', $all_cart_items->pluck('smartphone_id')->toArray())
+                    ->get();
+
+                $total_summary = $this->price_calculator->calculate($all_cart_items, $addon_cart_items);
+
+                return [
+                    'status' => true,
+                    'message' => $this->trans->get('Removed Succesfully'),
+                    'total_summary' => $total_summary,
+                ];
             }
 
-            $deleted = $item->delete();
-            if (! $deleted) {
-                throw new Exception($this->trans->get('Something Went Wrong While Removing Addon From Cart'));
+            $sessionKey = 'single_product_checkout_'.$user->id;
+            $sessionData = session()->get($sessionKey);
+
+            if (empty($sessionData) || empty($sessionData['addon_items'])) {
+                throw new Exception('Invalid Buy Now Session');
             }
 
-            $all_cart_items = $this->cart->where('customer_id', $customer->id)->with(['smartphone', 'smartphone.selling_info', 'smartphone.selling_info.shipping_fee', 'smartphone.selling_info.import_tax'])->get();
+            $addonItems = $sessionData['addon_items'];
+            $cartItems = $sessionData['cart_items'];
 
-            $addon_cart_items = $this->smartphone_cart_addon
-                ->where('customer_id', $customer->id)
-                ->whereIn('smartphone_id', $all_cart_items->pluck('smartphone_id')->toArray())
-                ->get();
+            $addonItems = $addonItems
+                ->reject(fn ($addon) => $addon->addon_id == $item_id)
+                ->values();
 
-            $total_summary = $this->price_calculator->calculate($all_cart_items, $addon_cart_items);
+            foreach ($cartItems as $cartItem) {
+                if ($cartItem->relationLoaded('smartphoneAddonItems')) {
+                    $cartItem->setRelation(
+                        'smartphoneAddonItems',
+                        $cartItem->smartphoneAddonItems
+                            ->reject(fn ($addon) => $addon->addon_id == $item_id)
+                            ->values()
+                    );
+                }
+            }
+
+            // recalculate totals
+            $total_summary = $this->price_calculator->calculate(
+                $sessionData['cart_items'],
+                $addonItems
+            );
+
+            // save back to session
+            session()->put($sessionKey, [
+                ...$sessionData,
+                'addon_items' => $addonItems,
+                'total_summary' => $total_summary,
+            ]);
 
             return [
                 'status' => true,
-                'message' => $this->trans->get('Removed Succesfully From Cart'),
+                'message' => 'Addon Removed Successfully',
                 'total_summary' => $total_summary,
             ];
+
         } catch (Exception $e) {
             return [
                 'status' => false,
@@ -768,6 +929,8 @@ class CartRepository implements ICartRepository
 
                 $addonItem->setRelation('addon', $addonModel);
                 $addonItems->push($addonItem);
+
+                $cartItem->setRelation('smartphoneAddonItems', $addonItems);
             }
 
             $total_summary = $this->price_calculator->calculate(
